tratamento de exceções

+++ src/main/java/com/example/meu_primeiro_springboot/service/ProdutoService.java
@@ -3,6 +3,7 @@ package com.example.meu_primeiro_springboot.service;
 import java.util.List;
 import java.util.Optional;
 
+import com.example.meu_primeiro_springboot.exceptions.RecursoNaoEncontradoException;
 import org.springframework.stereotype.Service;
 
 import com.example.meu_primeiro_springboot.model.Produto;
@@ -17,13 +18,17 @@ public class ProdutoService {
     public List<Produto> listarProdutos(){
         return produtoRepository.findAll();
     }
-    public Optional<Produto> buscarPorId(Long id){
-        return produtoRepository.findById(id);
+    public Produto buscarPorId(Long id){
+        return produtoRepository.findById(id)
+                .orElseThrow(() -> new RecursoNaoEncontradoException("Produto com" + id +"Não encontrado"));
     }
     public Produto salvarProduto(Produto produto){
         return  produtoRepository.save(produto);
     }
     public void deletarProduto(Long id){
+        if(!produtoRepository.existsById(id)){
+            throw new RecursoNaoEncontradoException("Produto com" + id +"Não encontrado");
+        }
         produtoRepository.deleteById(id);
     }
